requirements.txt
# Core ML/DL libraries
tensorflow>=2.13.0
torch>=2.0.0
torchvision>=0.15.0
transformers>=4.30.0
datasets>=2.14.0
evaluate>=0.4.0
scikit-learn>=1.3.0

# Data handling and visualization
numpy>=1.24.0
pandas>=2.0.0
matplotlib>=3.7.0
seaborn>=0.12.0
plotly>=5.15.0
Pillow>=10.0.0

# Web interface
streamlit>=1.25.0

# Configuration and utilities
PyYAML>=6.0
pathlib2>=2.3.7

# Development and testing
pytest>=7.4.0
pytest-cov>=4.1.0
black>=23.7.0
flake8>=6.0.0
mypy>=1.5.0

# Optional: For advanced features
opencv-python>=4.8.0
albumentations>=1.3.0
tensorboard>=2.13.0
wandb>=0.15.0
